- fetch: 优化几个工具文件 1. my_log 支持日志等级为INF时可以用MY_LOG_INF_DUMP输出DUMP日志，一些必要的场合需要，尤其是调试协议时用； 2. my_safe_memory简化，只保留几个核心api和相关的核心宏； 3. my_tool一些注释清理，有些注释指向了对应的模块，这不合理，工具为通用，不需要指向对应模块。

+++ project/OC810/code/utility/my_safe_memory.c
@@ -31,16 +31,16 @@ void my_safe_memory_check_health(void)
 
     if (freeHeap < MY_SAFE_MEM_CRITICAL_THRESHOLD)
     {
-        MY_LOG_E("[MEM] Low heap: %d bytes (%d%% used)", freeHeap, usage);
+        MY_LOG_E("[MEM] Low heap: %u bytes (%u%% used)", (unsigned)freeHeap, (unsigned)usage);
     }
     else if (freeHeap < MY_SAFE_MEM_WARN_THRESHOLD)
     {
-        MY_LOG_W("[MEM] Heap warning: %d bytes (%d%% used)", freeHeap, usage);
+        MY_LOG_W("[MEM] Heap warning: %u bytes (%u%% used)", (unsigned)freeHeap, (unsigned)usage);
     }
 
     if (minHeap < MY_SAFE_MEM_MIN_THRESHOLD)
     {
-        MY_LOG_E("[MEM] Critical min heap: %d bytes", minHeap);
+        MY_LOG_E("[MEM] Critical min heap: %u bytes", (unsigned)minHeap);
     }
 }
